+++ RealmSwift/Object.swift
@@ -49,10 +49,11 @@ import Realm.Private
  - `@objc enum` which has been delcared as conforming to `RealmEnum`.
  - `RealmOptional<Value>` for optional numeric properties
  - `Object` subclasses, to model many-to-one relationships
+ - `EmbeddedObject` subclasses, to model owning one-to-one relationships
  - `List<Element>`, to model many-to-many relationships
 
- `String`, `NSString`, `Date`, `NSDate`, `Data`, `NSData`, `Decimal128`, `ObjectId`  and
- `Object` subclass properties can be declared as optional.
+ `String`, `NSString`, `Date`, `NSDate`, `Data`, `NSData`, `Decimal128`, and `ObjectId`  properties
+ can be declared as optional. `Object` and `EmbeddedObject` subclasses *must* be declared as optional.
  `Int`, `Int8`, `Int16`, `Int32`, `Int64`, `Float`, `Double`, `Bool`,  enum, and `List` properties cannot.
  To store an optional number, use `RealmOptional<Int>`, `RealmOptional<Float>`, `RealmOptional<Double>`, or
  `RealmOptional<Bool>` instead, which wraps an optional numeric value. Lists cannot be optional at all.
@@ -140,8 +141,8 @@ open class Object: RLMObjectBase, RealmCollectionValue {
      It is not considered part of the public API.
      :nodoc:
      */
-    public override final class func _getProperties(withInstance instance: Any) -> [RLMProperty] {
-        return ObjectUtil.getSwiftProperties(instance as! RLMObjectBase)
+    public override final class func _getProperties() -> [RLMProperty] {
+        return ObjectUtil.getSwiftProperties(self)
     }
 
     // MARK: Object Customization
@@ -245,7 +246,7 @@ open class Object: RLMObjectBase, RealmCollectionValue {
      - returns: A token which must be held for as long as you want updates to be delivered.
      */
     public func observe(_ block: @escaping (ObjectChange) -> Void) -> NotificationToken {
-        return RLMObjectAddNotificationBlock(self, { names, oldValues, newValues, error in
+        return RLMObjectBaseAddNotificationBlock(self, { names, oldValues, newValues, error in
             if let error = error {
                 block(.error(error as NSError))
                 return
@@ -611,6 +612,15 @@ extension Object: _ManagedPropertyType {
 }
 
 /// :nodoc:
+extension EmbeddedObject: _ManagedPropertyType {
+    // swiftlint:disable:next identifier_name
+    public static func _rlmProperty(_ prop: RLMProperty) {
+        Object._rlmProperty(prop)
+        prop.objectClassName = className()
+    }
+}
+
+/// :nodoc:
 extension List: _ManagedPropertyType where Element: _ManagedPropertyType {
     // swiftlint:disable:next identifier_name
     public static func _rlmProperty(_ prop: RLMProperty) {
@@ -622,10 +632,10 @@ extension List: _ManagedPropertyType where Element: _ManagedPropertyType {
 }
 
 /// :nodoc:
-class LinkingObjectsAccessor<Element: Object>: RLMManagedPropertyAccessor {
+class LinkingObjectsAccessor<Element: ObjectBase>: RLMManagedPropertyAccessor where Element: RealmCollectionValue {
     @objc override class func initializeObject(_ ptr: UnsafeMutableRawPointer,
                                                parent: RLMObjectBase, property: RLMProperty) {
-        ptr.assumingMemoryBound(to: LinkingObjects.self).pointee.handle = RLMLinkingObjectsHandle(object: parent, property: property)
+        ptr.assumingMemoryBound(to: LinkingObjects<Element>.self).pointee.handle = RLMLinkingObjectsHandle(object: parent, property: property)
     }
     @objc override class func get(_ ptr: UnsafeMutableRawPointer) -> Any {
         return ptr.assumingMemoryBound(to: LinkingObjects<Element>.self).pointee
@@ -733,10 +743,10 @@ internal class ObjectUtil {
         }
     }
 
-    internal class func getSwiftProperties(_ object: RLMObjectBase) -> [RLMProperty] {
+    internal class func getSwiftProperties(_ cls: RLMObjectBase.Type) -> [RLMProperty] {
         _ = ObjectUtil.runOnce
 
-        let cls = type(of: object)
+        let object = cls.init()
 
         var indexedProperties: Set<String>!
         let columnNames = cls._realmColumnNames()